+++ R/standardize.R
@@ -5,63 +5,54 @@ utils::globalVariables(c("."))
 #' Perform data standardization for multivariate exploratory data analysis.
 #'
 #' @param X matrix
-#' @param type nature of standardization
-#' @param ... additional arguments
+#' @param center centering by the mean
+#' @param scale scaling by the standard deviation
+#' @param row.w an optional row weights (by default, a vector of 1 for uniform row weights); the weights are given only for the active individuals
 #'
 #' @details
-#' Standardization depends on what you need to perform factor analysis. Actually,
-#' there are only two types:
+#' Standardization depends on what you need to perform factor analysis. We
+#' implemented two types:
 #'
-#' * \code{type = "FactoMineR"}: standardization is performed by centering the
+#' * \code{standardize}: standardization is performed by centering the
 #' data matrix and dividing by the square root of the sum of squares of the
 #' weights. This is the default standardization in FactoMineR::PCA(..., scale.unit = TRUE).
 #'
-#' * \code{type = "norm"}: standardization is performed by centering and scaling
-#' the data matrix. This is the default standardization in R.
+#' * \code{standardize_norm}: standardization is performed by centering and scaling
+#' the data matrix. (X - µ) / σ, where µ is the mean and σ is the standard deviation.
 #'
 #' @examples
 #' library(FactoMineR2)
 #'
 #' iris[, -5] |>
-#'   standardize(type = "norm")
+#'   standardize_norm()
 #'
 #' @export
-standardize <- function(X, type = c("FactoMineR", "norm"), ...) {
-  type <- match.arg(type)
-  args <- list(...)
+standardize_norm <- function(X, center = TRUE, scale = TRUE) {
+  X <- scale(X, center = center, scale = scale)
+  return(X)
+}
 
-  if ("scale.unit" %in% names(args)) {
-    if (!is.logical(args$scale.unit)) {
-      stop("The argument `scale.unit` must be a logical value.")
-    }
-  } else {
-    args$scale.unit <- TRUE
+#' @rdname standardize_norm
+#' @export
+standardize <- function(X, scale = TRUE, row.w = NULL) {
+  if (is.null(row.w)) {
+    row.w <- rep(1, nrow(X))
   }
 
-  if (type == "norm") {
-    X <- scale(X, center = TRUE, scale = TRUE)
-  }
+  weights <- row.w / sum(row.w)
 
-  if (type == "FactoMineR") {
-    weights <- rep(1, times = nrow(X)) / sum(rep(1, times = nrow(X)))
+  center <- weights %>%
+    `/`(sum(weights)) %>%
+    crossprod(as.matrix(X)) %>%
+    as.vector()
 
-    center <- weights %>%
-      crossprod(as.matrix(X)) %>%
-      as.vector()
+  X <- t(t(as.matrix(X)) - center)
 
-    X <- t(t(as.matrix(X)) - center)
 
-    if (args$scale.unit == TRUE) {
-      std <- X^2 %>%
-        as.matrix() %>%
-        crossprod(weights / sum(weights), .) %>%
-        as.vector() %>%
-        sqrt()
-
-      std[std <= 1e-16] <- 1
-
-      X <- t(t(X) / std)
-    }
+  if (scale) {
+    std <- sqrt(as.vector(crossprod(weights / sum(weights), as.matrix(X^2))))
+    std[std <= 1e-16] <- 1
+    X <- t(t(X) / std)
   }
 
   return(X)